multi: make transport messages in-memory only
