+++ baselib/actor/AGENTS.md
@@ -2,15 +2,14 @@
 
 ## Purpose
 
-Core actor framework providing typed, message-driven concurrent components with
-durable mailbox persistence, service discovery via `Receptionist`, and
-crash-safe at-least-once delivery with exactly-once deduplication.
+Core actor framework providing typed, message-driven in-memory concurrent
+components with service discovery via `Receptionist`.
 
 ## Key Types
 
 - `Actor[M, R]` — Generic actor with typed message `M` and response `R`. Processes messages sequentially from its mailbox.
 - `ActorBehavior[M, R]` — Interface that actors implement: `Start`, `Receive`, `Stop`.
-- `ActorConfig[M, R]` — Configuration for actor creation (behavior, mailbox, codec, delivery store).
+- `ActorConfig[M, R]` — Configuration for actor creation (behavior, mailbox size, wait group, logger, and dead-letter outlet).
 - `ActorRef[M, R]` — Typed reference for sending messages to an actor (`Tell`, `Ask`).
 - `TellOnlyRef[M]` — Fire-and-forget reference (no response type).
 - `ActorSystem` — Container managing actor lifecycles, registration, and
@@ -25,15 +24,8 @@ crash-safe at-least-once delivery with exactly-once deduplication.
   remove all refs for this key.
 - `Receptionist` — Service locator mapping `ServiceKey` → `ActorRef` for decoupled actor wiring.
 - `Message` — Sealed interface for all actor messages (must embed `BaseMessage`).
-- `MessageCodec` — TLV-based codec for message serialization/deserialization.
-- `DeliveryStore` / `TxAwareDeliveryStore` — Interfaces for durable mailbox persistence (enqueue, claim, ack, dead-letter).
-- `DurableActor` — Actor variant with crash-safe mailbox backed by SQL persistence. Provides `Wait(ctx)` to block until the actor stops and `StopAndWait(ctx)` to request a graceful shutdown and then wait.
-- `DurableActorConfig[M, R]` — Configuration struct for `DurableActor`: behavior, store, codec, clock, DLO, WaitGroup, `TellRetryPolicy`, lease/heartbeat/poll durations, max attempts, cleanup timeout, and deduplication TTL.
-- `DefaultDurableActorConfig[M, R]()` — Constructor returning a `DurableActorConfig` with safe defaults (30s lease, 10 max attempts, 1s poll fallback, DefaultTellRetryPolicy).
 - `TellRetryPolicy` — Function type `func(attempts int, lastErr error) (bool, time.Duration)` determining retry behavior for failed Tell messages. Return `(false, _)` to dead-letter immediately.
 - `DefaultTellRetryPolicy` — Exponential backoff policy: up to 5 attempts, starting at 1s, capped at 60s.
-- `Checkpoint` — Serializable actor state snapshot for recovery.
-- `WithoutOutboxID` — Context helper that strips the propagated outbox ID so child operations do not inherit the parent's delivery tracking scope.
 - `Promise[T]` / `Future[T]` — Async result types for Ask-pattern responses.
 - `ChannelMailbox[M, R]` — In-memory channel-based mailbox (non-durable, for lightweight actors).
 - `Mailbox[M, R]` — Interface for actor message queues: `Send(ctx, env) error` (blocking, returns `ErrMailboxClosed` or `ErrActorTerminated` on failure), `TrySend(env) error` (non-blocking), `Receive(ctx) iter.Seq[envelope]`, `Close()`, `IsClosed() bool`, `Drain() iter.Seq[envelope]`. `Send` previously returned `bool`; it now propagates the exact failure cause so callers can distinguish `ErrMailboxClosed`, `ErrActorTerminated`, and `context.Canceled`/`context.DeadlineExceeded` without inspecting actor state.
@@ -41,23 +33,21 @@ crash-safe at-least-once delivery with exactly-once deduplication.
 
 ## Relationships
 
-- **Depends on**: `lnd/tlv` (message serialization).
-- **Depended on by**: All domain actors (`round`, `vtxo`, `oor`, `wallet`, `serverconn`, `timeout`, `indexer`), `baselib/protofsm` (FSM-to-actor bridge), `db/actordelivery` (persistence implementation).
+- **Depends on**: Go channels and context primitives.
+- **Depended on by**: All domain actors (`round`, `vtxo`, `oor`, `wallet`, `serverconn`, `timeout`, `indexer`) and `baselib/protofsm` (FSM-to-actor bridge).
 
 ## Invariants
 
 - Messages are processed sequentially per actor — no concurrent `Receive` calls.
-- `Tell` with a `DurableActor` persists the message before returning (crash-safe enqueue).
-- Outbox messages are dispatched only after state is persisted (outbox pattern).
+- `Tell` and `Ask` enqueue in memory; restart safety belongs to domain SQL tables and transport stores.
+- Outbox messages are dispatched by domain-owned SQL outboxes, not by actor message serialization.
 - `ServiceKey` lookup via `Receptionist` is type-safe: mismatched types return `ErrServiceKeyTypeMismatch`.
-- `RestartMessage` has `RestartPriority` (MaxInt32) ensuring it is processed before all other messages on recovery.
 - Transaction context (`WithTx`/`RequireTx`) enables same-DB-transaction joining between actors and their callers.
 - `Mailbox.Send` returns the exact failure error (`ErrMailboxClosed`, `ErrActorTerminated`, `context.Canceled`, `context.DeadlineExceeded`) rather than a boolean; `Tell` and `Ask` propagate this directly to callers.
-- During daemon teardown, the underlying DB is closed before every actor's lease loop has wound down. The lease loop uses `isExpectedShutdownErr` to demote these "database is closed" errors to debug level; real operational errors still surface as warnings because neither the actor context nor the outer context is done in those cases.
+- During daemon teardown, actors stop by closing their in-memory mailboxes and
+  waiting for behavior shutdown.
 
 ## Deep Docs
 
 - [baselib/CLAUDE.md](../CLAUDE.md) — Parent baselib package overview.
-- [docs/durable_actor_architecture.md](../../docs/durable_actor_architecture.md) — Durable actor internals.
-- [docs/durable_actor_quickstart.md](../../docs/durable_actor_quickstart.md) — TLVMessage, ActorBehavior, migration checklist.
 - [ARCHITECTURE.md](../../ARCHITECTURE.md) — System-wide package map.
